added the authenication and authorization to the features controller

+++ controller/feature.controller.js
@@ -1,16 +1,34 @@
 const { Goal, Bill, Statement } = require("../models/features.model");
+const { User } = require("../models/account.model");
 const errorHandler = require("../middleware/errorHandler");
+const jwt = require("jsonwebtoken");
+
+const bcrypt = require("bcryptjs");
+require("dotenv").config;
 
 class BillController {
   async createBill(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+
+      const userId = decodedToken._id;
+
       const { account, payee, amount, dueDate } = req.body;
       const newBill = new Bill({
         account,
         payee,
         amount,
         dueDate,
+        userId,
       });
+
       await newBill.save();
       res.status(201).json(newBill);
     } catch (error) {
@@ -18,7 +36,17 @@ class BillController {
     }
   }
   async getBill(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+      const token = authorization.split(" ")[1];
+      const decodeToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodeToken._id;
+
       const billId = req.params.billId;
 
       const bill = await Bill.findById(billId);
@@ -28,40 +56,87 @@ class BillController {
           message: "Bill not found",
         });
       }
+      if (userId.toString() !== bill.userId.toString()) {
+        return res.status(403).json({
+          message:
+            "Forbidden: You do not have permission to access this account.",
+        });
+      }
       res.json(bill);
     } catch (error) {
       errorHandler(error, req, res);
     }
   }
   async updateBill(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodedToken._id;
+
       const billId = req.params.billId;
+
+      const bill = await Bill.findById(billId);
+
       const { account, payee, amount, dueDate, isPaid } = req.body;
 
-      const updatedBill = await Bill.findByIdAndUpdate(
-        billId,
-        {
-          account,
-          payee,
-          amount,
-          dueDate,
-          isPaid,
-        },
-        { new: true }
-      );
-      if (!updatedBill) {
+      if (!bill) {
         return res.status(404).json({
           message: "Bill not found",
         });
       }
+
+      if (userId.toString() !== bill.userId.toString()) {
+        return res.status(403).json({
+          message: "Forbidden: You do not have permission to access this bill.",
+        });
+      }
+
+      bill.account = account;
+      bill.payee = payee;
+      bill.amount = amount;
+      bill.dueDate = dueDate;
+      bill.isPaid = isPaid;
+
+      await bill.save();
     } catch (error) {
       errorHandler(error, req, res);
     }
   }
 
   async deleteBill(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
       const billId = req.params.billId;
+
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodedToken._id;
+
+      const bill = await Bill.findById(billId);
+
+      if (!bill) {
+        return res.status(404).json({
+          message: "Bill not found",
+        });
+      }
+
+      if (userId.toString() !== bill.userId.toString()) {
+        return res.status(403).json({
+          message: "Forbidden: You do not have permission to delete this bill.",
+        });
+      }
 
       const deleteBill = await Bill.findByIdAndRemove(billId);
 
@@ -79,12 +154,25 @@ class BillController {
 
 class StatementController {
   async generateStatement(req, res) {
+    const { authorization } = req.headers;
+
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+
+      const userId = decodedToken._id;
+
       const { account, startDate, endDate } = req.body;
       const newStatement = new Statement({
         account,
         startDate,
         endDate,
+        userId,
       });
       await newStatement.save();
       res.status(201).json(newStatement);
@@ -94,7 +182,17 @@ class StatementController {
   }
 
   async getStatement(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+      const token = authorization.split(" ")[1];
+      const decodeToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodeToken._id;
+
       const statementId = req.params.statementId;
 
       const statement = await Statement.findById(statementId).populate(
@@ -103,6 +201,12 @@ class StatementController {
       if (!statement) {
         return res.status(404).json({
           message: "Statement not found",
+        });
+      }
+      if (userId.toString() !== statement.statementId.toString()) {
+        return res.status(403).json({
+          message:
+            "Forbidden: You do not have permission to access this account.",
         });
       }
       res.json(statement);
@@ -114,7 +218,17 @@ class StatementController {
 
 class GoalController {
   async createGoal(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodedToken._id;
+
       const { account, goalName, targetAmount, currentAmount, deadline } =
         req.body;
 
@@ -124,6 +238,7 @@ class GoalController {
         targetAmount,
         currentAmount,
         deadline,
+        userId,
       });
       await newGoal.save();
 
@@ -134,16 +249,33 @@ class GoalController {
   }
 
   async getGoal(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodedToken._id;
+
       const goalId = req.params.goalId;
 
       const goal = await Goal.findById(goalId);
 
       if (!goal) {
         return res.status(404).json({
-          message: "Goal not found ",
+          message: "Goal not found",
         });
       }
+
+      if (userId.toString() !== goal.userId.toString()) {
+        return res.status(403).json({
+          message: "Forbidden: You do not have permission to access this goal.",
+        });
+      }
+
       res.json(goal);
     } catch (error) {
       errorHandler(error, req, res);
@@ -151,38 +283,80 @@ class GoalController {
   }
 
   async updateGoal(req, res) {
+    const { authorization } = req.headers;
     try {
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
+
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodedToken._id;
+
       const goalId = req.params.goalId;
+
+      const goal = await Goal.findById(goalId);
+
       const { goalName, targetAmount, currentAmount, deadline } = req.body;
-      const updateGoal = await Goal.findByIdAndUpdate(
-        goalId,
-        {
-          goalName,
-          targetAmount,
-          currentAmount,
-          deadline,
-        },
-        { new: true }
-      );
-      if (!updateGoal) {
+
+      if (!goal) {
         return res.status(404).json({
           message: "Goal not found",
         });
       }
+
+      if (userId.toString() !== goal.userId.toString()) {
+        return res.status(403).json({
+          message: "Forbidden: You do not have permission to update this goal.",
+        });
+      }
+      goal.goalName = goalName;
+      goal.targetAmount = targetAmount;
+      goal.currentAmount = currentAmount;
+      goal.deadline = deadline;
+      await goal.save();
+      res.json(goal);
     } catch (error) {
       errorHandler(error, req, res);
     }
   }
 
   async deleteGoal(req, res) {
+    const { authorization } = req.headers;
     try {
-      const gaolId = req.params.goalId;
+      if (!authorization) {
+        return res.status(401).json({
+          error: "Unauthorized: Missing Authorization Header",
+        });
+      }
 
-      const deletedGoal = await Goal.findByIdAndRemove(gaolId);
+      const goalId = req.params.goalId;
+
+      const token = authorization.split(" ")[1];
+      const decodedToken = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+      const userId = decodedToken._id;
+
+      const goal = await Goal.findById(goalId);
+
+      if (!goal) {
+        return res.status(404).json({
+          message: "Goal not found",
+        });
+      }
+
+      if (userId.toString() !== goal.userId.toString()) {
+        return res.status(403).json({
+          message: "Forbidden: You do not have permission to delete this goal.",
+        });
+      }
+
+      const deletedGoal = await Goal.findByIdAndRemove(goalId);
 
       if (!deletedGoal) {
         return res.status(404).json({
-          message: "gaol not found",
+          message: "Goal not found",
         });
       }
       res.json(deletedGoal);
@@ -191,4 +365,5 @@ class GoalController {
     }
   }
 }
+
 module.exports = { BillController, GoalController, StatementController };
